Change input from flat list to nested list
Using a 2-D list, with each inner list representing a row of squares on the
game board, makes for easier math when referencing individual squares (i.e.
don't need to convert from list index to x, y coords.).

The width of the board is also contained in the board itself
(i.e. simply the length of the first row), and does not have to be passed
in as a separate parameter.

Printing the output is also be made simpler (simply print one row at a
time).