Read input from .csv, automate output using curses
The starting configuration (seed) of squares can be read from a .csv file.
The file name is passed to the program with the -f flag.
A few simple sample seeds are provided in the samples directory.
This makes it easier to input seeds, and in a reusable way.

The output of each iteration of the game is printed to the screen with
the help of curses, so the board can be printed to the same space, which
looks prettier and makes it easier to see the evolution of the squares.